Finalizacion de los servicios de localizacion que se requieren por los momentos.

+++ src/main/java/com/okiimport/app/service/web/SLocalizacion.java
@@ -4,6 +4,6 @@ import com.okiimport.app.model.Ciudad;
 import com.okiimport.app.model.Pais;
 
 public interface SLocalizacion {
-	long calcularDistancia(Ciudad ciudadOrigen, Ciudad ciudadDestino);
-	long calcularDistancia(Pais paisOrigen, Pais paisDestino);
+	double calcularDistancia(Ciudad ciudadOrigen, Ciudad ciudadDestino);
+	double calcularDistancia(Pais paisOrigen, Pais paisDestino);
 }
